requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-storage==2.13.0
python-dotenv==1.0.0
pydantic==2.5.2
anthropic==0.8.1
numpy==1.26.2
open3d==0.17.0 